channelmanager: Switch to new Channel Manager.
Reuses components from old Admin Channel manager,
now used universally for all users.
