+++ ui/src/channels/ChannelHeader.tsx
@@ -21,8 +21,8 @@ import {
 import { useChatState } from '@/state/chat';
 import { useDiaryState } from '@/state/diary';
 import { useHeapState } from '@/state/heap/heap';
-import EditChannelModal from '@/groups/GroupAdmin/AdminChannels/EditChannelModal';
-import DeleteChannelModal from '@/groups/GroupAdmin/AdminChannels/DeleteChannelModal';
+import EditChannelModal from '@/groups/ChannelsList/EditChannelModal';
+import DeleteChannelModal from '@/groups/ChannelsList/DeleteChannelModal';
 import CaretLeftIcon from '@/components/icons/CaretLeftIcon';
 import ChannelIcon from '@/channels/ChannelIcon';
 import Divider from '@/components/Divider';
@@ -121,7 +121,7 @@ function ChannelActions({
       );
     } catch (error) {
       if (error) {
-        console.error(`[ChannelIndex:LeaveError] ${error}`);
+        console.error(`[ChannelHeader:LeaveError] ${error}`);
       }
     }
   }, [flag, ship, name, navigate, leave, isMobile]);
@@ -185,6 +185,7 @@ function ChannelActions({
             setEditIsOpen={setEditIsOpen}
             nest={nest}
             channel={channel}
+            app={_app}
           />
           <DeleteChannelModal
             deleteChannelIsOpen={deleteChannelIsOpen}
