+++ ui/src/groups/ChannelsList/ChannelPermsSelector.tsx
